Revert "[QFEED-41] Refactor: chatmessage 단일 도메인 반환 dto 생성"
This reverts commit 179be53a1040fdfbeaa4e55e8b95a1546516984d.
