+++ module-domain/src/main/java/com/wsws/moduledomain/chat/ChatMessage.java
@@ -19,26 +19,26 @@ public class ChatMessage {
     private Boolean isRead = false;
     private LocalDateTime createdAt;
     private UserId userId;
-    private Long chatRoomId;
+    private ChatRoom chatRoom;
 
     // 생성자
-    private ChatMessage(Content content, MessageType type, String url, Boolean isRead, LocalDateTime createdAt, UserId userId, Long chatRoomId) {
+    private ChatMessage(Content content, MessageType type, String url, Boolean isRead, LocalDateTime createdAt, UserId userId, ChatRoom chatRoom) {
         this.content = content;
         this.type = type;
         this.url = url;
         this.isRead = isRead;
         this.createdAt = createdAt;
         this.userId = userId;
-        this.chatRoomId = chatRoomId;
+        this.chatRoom = chatRoom;
     }
 
     // 팩토리 메서드
-    public static ChatMessage create(String content, MessageType type, String url, UserId userId, Long chatRoomId) {
+    public static ChatMessage create(String content, MessageType type, String url, UserId userId, ChatRoom chatRoom) {
         if (type != MessageType.TEXT && (url == null || url.isEmpty())) {
             throw new UrlRequiredException(); // 예외는 직접 정의해야 함
         }
         Content chatContent = Content.from(content);
-        return new ChatMessage(chatContent, type, url, false, LocalDateTime.now(), userId, chatRoomId);
+        return new ChatMessage(chatContent, type, url, false, LocalDateTime.now(), userId, chatRoom);
     }
 
     // 읽음 처리 메서드
