+++ module-domain/src/main/java/com/wsws/moduledomain/chat/repo/ChatMessageRepository.java
@@ -1,11 +1,8 @@
 package com.wsws.moduledomain.chat.repo;
 
 import com.wsws.moduledomain.chat.ChatMessage;
-import com.wsws.moduledomain.chat.vo.ChatMessageUserInfraDTO;
-import com.wsws.moduledomain.chat.vo.ChatMessageInfraDto;
-import com.wsws.moduledomain.chat.vo.ChatMessageUserInfraDTO;
-import com.wsws.moduledomain.chat.vo.ChatRoomInfraDto;
-
+import com.wsws.moduledomain.chat.ChatRoom;
+import com.wsws.moduledomain.chat.dto.ChatMessageDTO;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,9 +11,9 @@ public interface ChatMessageRepository {
 
     void markAllMessagesAsRead(Long chatRoomId);
 
-    List<ChatMessageUserInfraDTO> findMessagesWithUserDetails(Long chatRoomId, int page, int size);
+    List<ChatMessageDTO> findMessagesWithUserDetails(Long chatRoomId, int page, int size);
     // 특정 채팅방에서 마지막 메시지를 가져오기 (최신 메시지 하나만)
-    Optional<ChatMessageInfraDto> findTopByChatRoomOrderByCreatedAtDesc(ChatRoomInfraDto chatRoom);
+    Optional<ChatMessage> findTopByChatRoomOrderByCreatedAtDesc(ChatRoom chatRoom);
 
     long countUnreadMessages(Long chatRoomId,String userId);
 }
